[AutoComplete] Add testing code to test Trie structure
Various code were added in various places to test the Trie structure.
This structure is used for the AutoComplete feature

As of now, pressing Tab will invoke a method from the Trie object to print all names in the addressbook
This printing is done by pulling the names from the Trie structure itself.

The testing code are TEMPORARY