Fix text prediction appending space bug
The bug was given that the command box already contains the ONLY
possible predicted value, invoking text prediction will append
another single whitespace character even though a whitespace character
is already appended at the end.

This behaviour is fixed to prevent appending of whitespace character
unnecessarily. The fix is achieved by deleting a unnecessary line in
getPredictList(String) in Trie.java where the line adds a whitespace
when the end of a branch in the Trie structure is detected.