Fix index out of bounds bug
Bug:
Invoking text prediction on a searching prefix with a single character
will cause an index out of bound error.

Problem:
The algorithm to build the string to be appended to the predictions
list will delete character by character. Invoking this exploring algo
on a prefix with a single character will cause the algo to try to
delete a character even if the string is empty (size == 0).

Fix:
Added a conditional check to check if the string has size > 0 before
deleting a character.