Fix wrong word predictions bug
Bug:
When there are multiple word entries with similar prefixes to be
searched, some predicted texts are not accurate (the text displayed is
not the same as the value entered into the data structure).

Issue:
There was a condition to check and delete the last whitespace characters
appended to end of a word so the string of characters that is built
recursively, is built properly. However, this condition also checked and
deleted whitespace characters that were part of the entire predicted
string. Hence, the string of characters to be built recursively is not
being built properly.

Fix:
Refined the condtional check to check that the current node during the
traversal of the data structure is an end node of the branch. This
ensures that only whitespaces appended to the end of a word is being
deleted here.