GroupCommand now adds a group
Instead of throwing an exception, GroupCommand now adds a group to the addressbook.
The group will also be saved and written to the addressbook XML file.
 
New Classes Created : Group, UniqueGroupList, XMLAdaptedGroup, DuplicateGroupException

Note: When reading  from the XML file, the addressbook will not read groups as of yet.
JUnit tests not fully complete.