Fix text prediction appending space bug
Bug:
When invoking text prediction, having a white space character as the
last character in the user input, will result in more white space
character being appended at each invoke. Though it should list the
possible predictions, the behaviour does not.
It appends a white space instead.

Issue:
In the extractArgumentValue() method of ArgumentTokenizer class,
the return string value is trimmed. Text prediction uses this
method for parsing of user input. Hence when there is a single
white space at the end of the user input, that whitespace is trimmed.
What this means is that the value passed into the text prediction
classes for prediction, does not contains that white space.

Important thing to note here is, that white space is important
because the data structure containing the prediction strings takes
into account of white spaces (e.g. contact address values). Hence,
removing white space from the user input will cause text prediction
to predict incorrectly.

Fix:
Prevent extractArgumentValue() method to trim by default. Instead,
it only trims if the untrimmed string value contains only white
spaces. In that case, the method will trim it.
This way, text prediction will receive untrimmed user input and
work correctly. In addition, this also provides flexibility of input
for the user since accidentally input a white space will still allow
the command to be parsed and processed.