Fix text prediction not appending space bug
Bug:
If the user input to be predicted is already the completed text,
invoking text prediction should append a white space if there isn't.
However, space are not appended.

Fix:
Change the conditional check in line 180 of Trie.java to check for 1
child or less instead of exaclty 1 child. This will take into account
when the input to be predicted is already a complete word and append
space accordingly.