Add a TestCase that throws CommandException when a duplicate GroupName is found when DistributeCommand is ran.
Test Case include having a StubAddressBook with several Person in it.
I have added a Group with an expected duplicated name with the DistributeBuilder.
I have load the stubAddressBook into a stub Model and a default Distribute object by the DistributeBuilder and pass it into the DistributeCommand Constructor.
Predicts the expected Error Message by using thrown.expectedMessage
Expected Error is encounted.