Fix Undo command when DistributeCommand is called.
Add an boolean flag to enable or disable commit function.
Since AddGroupCommand and CreateGroupCommand are called when Distribute Command is executed. I can set the shouldCommit to be false for both AddGroupCommand and CreateGroupCommand. This allow me to not commit any changes until the DistributeCommand is completely executed. It is also necessary to set the shouldCommit variable back to true to not affect other operations.
Changed all methods in DistributeUtil to public for any class to access it's functionalities if required.