For new issue - Lowercase Commands and Commands with and without underscore.
All commands are now standardised all in small letters, and support both with underscore spacing and without - for words that are made up of 2 or more actual words.
Also updated the suggestions, UI will only suggest the non-underscore format, but if the user decides to still enter with the underscore, the command still works and is considered an actual command.