Fix and acknowledge bug
We have come to find that the parse only takes the last occurence of the prefix and its content. This will prove to be problematic for EmailCommand as the occurence of the prefix in the content of the subject or message is not uncommon.
For now, we will take the first occurence of the prefix and its content and included a note in the UserGuide to inform users to avoid having the prefix in the content.